requirements.txt
pyperclip
pynput
pycrypto